requirements.txt
matplotlib==3.5.3
numpy==1.21.5
pandas==2.0.1
PyYAML==6.0.1
scikit_learn==1.3.0
torch==1.13.0
xgboost==1.7.5
